broke cube and wall into additional components as part of logo folder
